Implement `collections.abc.Mapping` for `Multiaddr` to give it a dict-style feel
Nit: All operations on this “dictionary” (including __len__ and __getitem__) are O(n) over the number of
bytes/protocols of the MultiAddr, rather then using more optimized code that would require pre-parsing
the binary MultiAddr. Considering that MultiAddr tend to be rather short this should hopefully not be
an issue, but changing this would be possible in the future without changing any part of the interface.